backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-generativeai==0.7.0
supabase==2.5.0
python-dotenv==1.0.1
pydantic==2.7.0
httpx==0.27.0
websockets==12.0
numpy==1.26.4
scipy==1.13.0
exponent-server-sdk==2.1.0
pandas==2.2.0
psycopg2-binary==2.9.9
